Splitting tests by file and noting missing tests
