+++ qiita_pet/handlers/study_handlers/tests/test_description_handlers.py
@@ -1,0 +1,18 @@
+# -----------------------------------------------------------------------------
+# Copyright (c) 2014--, The Qiita Development Team.
+#
+# Distributed under the terms of the BSD 3-clause License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# -----------------------------------------------------------------------------
+from unittest import main
+
+from qiita_pet.test.tornado_test_base import TestHandlerBase
+
+
+class PreprocessingSummaryHandlerTests(TestHandlerBase):
+    # TODO: Missing tests
+    pass
+
+if __name__ == "__main__":
+    main()
