+++ qiita_pet/handlers/study_handlers/tests/test_edit_handlers.py
@@ -1,0 +1,140 @@
+# -*- coding: utf-8 -*-
+# -----------------------------------------------------------------------------
+# Copyright (c) 2014--, The Qiita Development Team.
+#
+# Distributed under the terms of the BSD 3-clause License.
+#
+# The full license is in the file LICENSE, distributed with this software.
+# -----------------------------------------------------------------------------
+from unittest import main
+
+from qiita_pet.test.tornado_test_base import TestHandlerBase
+from qiita_db.study import StudyPerson, Study
+from qiita_db.util import get_count, check_count
+
+
+class TestStudyEditorForm(TestHandlerBase):
+    # TODO: add proper test for this once figure out how. Issue 567
+    pass
+
+
+class TestStudyEditorExtendedForm(TestHandlerBase):
+    # TODO: add proper test for this once figure out how. Issue 567
+    pass
+
+
+class TestStudyEditHandler(TestHandlerBase):
+    database = True
+
+    def test_get(self):
+        """Make sure the page loads when no arguments are passed"""
+        response = self.get('/study/create/')
+        self.assertEqual(response.code, 200)
+        self.assertNotEqual(str(response.body), "")
+
+    def test_get_edit(self):
+        """Make sure the page loads when we want to edit a study"""
+        response = self.get('/study/edit/1')
+        self.assertEqual(response.code, 200)
+        self.assertNotEqual(str(response.body), "")
+
+    def test_get_edit_utf8(self):
+        """Make sure the page loads when utf8 characters are present"""
+        study = Study(1)
+        study.title = "TEST_ø"
+        study.alias = "TEST_ø"
+        study.description = "TEST_ø"
+        study.abstract = "TEST_ø"
+        response = self.get('/study/edit/1')
+        self.assertEqual(response.code, 200)
+        self.assertNotEqual(str(response.body), "")
+
+    def test_post(self):
+        person_count_before = get_count('qiita.study_person')
+        study_count_before = get_count('qiita.study')
+
+        post_data = {'new_people_names': ['Adam', 'Ethan'],
+                     'new_people_emails': ['a@mail.com', 'e@mail.com'],
+                     'new_people_affiliations': ['CU Boulder', 'NYU'],
+                     'new_people_addresses': ['Some St., Boulder, CO 80305',
+                                              ''],
+                     'new_people_phones': ['', ''],
+                     'study_title': 'dummy title',
+                     'study_alias': 'dummy alias',
+                     'pubmed_id': 'dummy pmid',
+                     'environmental_packages': ['air'],
+                     'timeseries': '1',
+                     'study_abstract': "dummy abstract",
+                     'study_description': 'dummy description',
+                     'principal_investigator': '-2',
+                     'lab_person': '1'}
+
+        self.post('/study/create/', post_data)
+
+        # Check that the new person was created
+        expected_id = person_count_before + 1
+        self.assertTrue(check_count('qiita.study_person', expected_id))
+
+        new_person = StudyPerson(expected_id)
+        self.assertTrue(new_person.name == 'Ethan')
+        self.assertTrue(new_person.email == 'e@mail.com')
+        self.assertTrue(new_person.affiliation == 'NYU')
+        self.assertTrue(new_person.address is None)
+        self.assertTrue(new_person.phone is None)
+
+        # Check the study was created
+        expected_id = study_count_before + 1
+        self.assertTrue(check_count('qiita.study', expected_id))
+
+    def test_post_edit(self):
+        study_count_before = get_count('qiita.study')
+        study = Study(1)
+        study_info = study.info
+
+        post_data = {
+            'new_people_names': [],
+            'new_people_emails': [],
+            'new_people_affiliations': [],
+            'new_people_addresses': [],
+            'new_people_phones': [],
+            'study_title': 'dummy title',
+            'study_alias': study_info['study_alias'],
+            'publications_doi': ','.join(
+                [doi for doi, _ in study.publications]),
+            'study_abstract': study_info['study_abstract'],
+            'study_description': study_info['study_description'],
+            'principal_investigator': study_info['principal_investigator'].id,
+            'lab_person': study_info['lab_person'].id}
+
+        self.post('/study/edit/1', post_data)
+
+        # Check that the study was updated
+        self.assertTrue(check_count('qiita.study', study_count_before))
+        self.assertEqual(study.title, 'dummy title')
+
+
+class TestCreateStudyAJAX(TestHandlerBase):
+    database = True
+
+    def test_get(self):
+
+        response = self.get('/check_study/', {'study_title': 'notreal'})
+        self.assertEqual(response.code, 200)
+        # make sure responds properly
+        self.assertEqual(response.body, 'True')
+
+        response = self.get('/check_study/')
+        self.assertEqual(response.code, 200)
+        # make sure responds properly
+        self.assertEqual(response.body, 'False')
+
+        response = self.get(
+            '/check_study/',
+            {'study_title':
+             'Identification of the Microbiomes for Cannabis Soils'})
+        self.assertEqual(response.code, 200)
+        # make sure responds properly
+        self.assertEqual(response.body, 'False')
+
+if __name__ == "__main__":
+    main()
